fix race condition blocking shutdown
the case is:
call CoreARI.run() in a thread
then quickly call CoreARI.stop() in another thread

This leads to the call CoreARI.run() to never stop, because the
websockets was not created at the time of CoreARI.stop(), which did
nothing. Then joining the run() thread would never return.