transfers: add the ability to set variables on a channel
The POST on channel now accepts variables that are used in the
originate. These variable are useful to do "special" transfers that
would require new URLs to implement.

for example, I can transfer to s@vmbox with XIVO_BASE_CONTEXT=default
ARG1=1001 to leave a message in the mailbox 1001 in context default.