add chats plugin
Add two new ressources:

* POST /chats
* POST /users/me/chats

The first one allow to send a chat message to an arbitrary user (on a
arbitrary xivo, no check is done to check if the destination exist) from
an arbitrary user (but only on the same xivo, altough no check is done
if the sender UUID really match an existing user). It requires an ACL
that is reserved to non-user token.

The second one is similar, but doesn't allow specifying the sender: the
user UUID is the user UUID of the token passed in.