get_token_user_uuid: handle xivo-auth errors
This doesn't happens often, since the get_token_user_uuid_from_request
function is called just after the ACL check is done. But this can still
happens (token expired, xivo-auth stopped, network problems, etc).

Still, the "best" solution to this problem would probably be to cache
the token used when the ACL validation is done, so that only 1 request to
xivo-auth is required to check the ACL + get the token user UUID,
instead of 2.