voicemails: add cache cleanup
This is a bit of a poor's man implementation, i.e. every X time the
get_diff method is called, we removed the voicemails for which there is
no directory on the file system. This ensure that the cache doesn't grow
in an unbound fashion, but it would be cleaner to react to a specific
event "the voicemail with number X and context Y has been removed", but
there is a few things missing before being able to do so.