calls: originate from mobile shows mobile callerid
reason: right now, the mobile and the destination have to see the same
callerid, because of the originate via Local channel. So we choose to
display the callerid of the mobile on both legs, so that the
destination knows who is calling. The mobile will see its own number
as callerid, but that is the lesser wrong, because the mobile call was
knowingly started by the user.