int tests: convert docker-compose v1 to v3
reason: we will add mongooseim and the container need to be linked by
ctid-ng container but also need to link ctid-ng container (circular
dependency).

In docker-compose v3, all dockers are in the same network. So the
"links" key are useless if we simply want communicate between
containers.

The "links" key is always necessary in the "sync" container, because it
start others containers.