int tests: readd 'links: ari' to fix tests
reason: to be sure that ari is started before ctid-ng, otherwise ctid-ng
fail to start. Since ari is a mock, it start instantly and ready to
listen when the container is started. But with a real ari (or longer
init), this method should not be enought.

We should probably try to reconnect several times on ari instead
of instantly fail.