chats: do not store remote messages in multi-wazo
MongooseIM does not archive a message if it doesn't own the domain.
Example:
caller: alice@localhost
to: bob@localhost
Then we store 2 messages; 1 for alice and 1 for bob
But if bob is in other wazo, we do not want to store his messages.
So if we put a different domain than localhost (e.g. to_xivo_uuid), then
MongooseIM will not kept the message in DB