add clean target to remove docker image
To avoid usage of cache. Example:
I update xivo-bus to change marshaler interface.
When daily-integration-tests is executed
Then docker use cache to build wazopbx/xivo-ctid-ng
Result: xivo-bus in the docker image is not the last version

Another solution should be to write the xivo-bus version in the
requirements.txt. But it would be painful to maintain.