relocates: test relocate from mobile to line
note: WAZO_DEREFERENCED_USERUUID is the XIVO_USERUUID of the non-local
channel generated by the mobile call. Calculating it for real would be
very costly, so we set it up front when originating the mobile
call. As a result, XIVO_USERUUID is not the only variable to contain
the channels user uuid.

However, XIVO_USERUUID should still be ignored on Local channels:

- XIVO_USERUUID on Local;2 will be used to dial the first leg (check
call permissions, etc.)
- XIVO_USERUUID on Local;1 will be used to dial the second leg (again,
check call permissions, specific trunks for specific users, etc.)

This means XIVO_USERUUID on a Local channel is different from
XIVO_USERUUID on non-Local channels: it was used to act on behalf of a
user for executing dialplan, but we can't set distinct values of
XIVO_USERUUID for Local;1 and Local;2, so it's not reliable enough to
interpret it as "who does this channel belong to".