tests: fix condition for existing calls
reason: the bus events is sent when the AMI event is received, which
happens a bit sooner than when the call is ready to be used in a
transfer, i.e. has an application_instance registered in the state
persisttor. In some cases (Jenkins), the transfer is started before
the call is ready to be used. This ensures the calls are registered in
the call state persistor.