calls: add the user uuid for calls originated from /applications
Calls originated from the calls API know the user which does the originate and
adds the required channel variables to the originate to be able to get that
info when doing a GET on /calls.

The /applications originate on the otherhand is user agnostic and uses the
context/extention combination to do the originate.

This patch uses another patch in xivo-agid which sets the WAZO_DIALED_USER_UUID
variable on the ;2 side of the local channel when going through the user
dialplan.