application: remove duplicate _creation_destinations
The 'self._create_destinations' in initialize was called before that the
application was created by the websocket. So the bridge was created but
not subscribed (error from ari in the service.py) and no event was sent.
Since the bridge was already created, it was not recreated by the
'self._on_start'.
So we reove the 'create_destinations' in initialize and wait before the
websocket, which trigger the 'self._on_start', which trigger the
'create_destinations'