transfers: remove dependency to calls plugin
reason:
Given the calls plugin is loaded after the transfers plugin
Given the initiator and transferred are not in stasis
When the transfer is started
When the initiator is the last call to be moved to stasis (last StasisStart)
Then the recipient does not ring

Because the StasisStart callbacks from calls can be executed after the
callbacks from transfers, the transfers plugin cannot depend on the
calls plugin.

app_instance does not have a clear usage, and the recipient probably
does not need to have the same app_instance than the initiator.