calls: add the sip_call_id field
the sip_call_id contains the value of the Call-ID SIP header. it is
used to match an event to a call started using a webrtc phone.

without this information the call initiator has to rely on the caller
id name and number of the call to "guess" if an event is related to its
phone. The guessing method also does not work with multiple calls.