call_ended: add the sip_call_id field
When a call is created the sip_call_id is assigned to the WAZO_SIP_CALL_ID
channel variable. Then when the hangup AMI event is received containing this
variable the call_ended event can be populated from this variable.

the channel helper sip_call_id function cannot be used at this point since the
channel has already disappeared