dial mobile: hang up all unanswered channels on answer
When calling a user with multiple contacts when the first one gets answered or
when the called hangs up all devices that did not answer the call must be hung
up.

If a device is not hung up, it will keep ringing.

Instead of hanging up all channels with a state "Ringing" this change hangs up
all channels that are not in state "Up".

One of the case I have observed is that the channel will have a state "Down"
before it starts ringing. In that case, that device will keep ringing after the
phone got answered elsewhere or got hung up by the caller.