trunk status: handle error on dissociation
If I dissociate my trunk from the SIP endpoint. Calld will receive 2 events. A
dissociation event from wazo-confd and a Registry event from Asterisk. When the
Registry events is received, the trunk has already been dissociated from the
endpoit and cannot be found.

If the the Status of the Registry event is Registered the exception will be
raised because there's something wrong