conference: optimize confd calls
Given the same conference ID avoid doing multiple queries on wazo-confd. This is
especially useful when a paging is used. In these scenario each members of the
paging will generate 2 confd calls which will then generate 1 auth call.

For a paging with 100 members this reduces the number of HTTP requests from 400
to 2.

The cache is reset whenever a conference is added, updated or deleted