switchboard: go back to the dialplan on stasis error
Calling continueInDialplan will allow the channel to follow through with the
dialplan and any error handling that could eventually happen there.

Without this error handling an exception is logged in the log file of wazo-calld
and the call is left hanging in the great void until the caller gets tired of
waiting and eventually hangs up or Asterisk gets restarted.