mobile push: set a custom hint on mobile sessions
this hint is defined by wazo-confgend. it is meant to be used to monitor the
state of the usersharedline extensions. The custom hint is set to NOT_INUSE when
a mobile sessions exists for a given user. This change means that a
usersharedline will be seen as available in a group or a queue if the user has a
mobile session. Allowing the user to receive calls from the group.