push: publish the cancel_notification event
this event is used to stop a mobile from ringing when it received a push mobile
event to receive a call.

When a push happens the call can be abandonned by the caller, answered by the
mobile or answered elsewhere.

If the call is abandonned or answered elsewhere the mobile needs to get notified
that it can stop ringing. In the best case scenario the mobile sent it's
REGISTER, received an INVITE and receives a BYE. But waking up a mobile and
registering it on Asterisk takes a long time and when the call gets abandonned
or answered before the REGISTER then the mobile has no way of knowing that it
should stop ringing.