transfer: check for answered transfers on start
using wazo-res-stasis-amqp allows the stasis app to be used even if wazo-calld
is stopped. This is because the stasis application is not bound to a websocket
anymore. This means that the recipient channel will not get an error when
entering the Stasis application when answering.

so when starting we are going to check the status of ongoing transfers to see if
a recipient call that was in ringback mode got answered while wazo-calld was down.