make a collectd test less flaky
calld needs to write its state variable for the new call before being restarted
for the expected processing to happen. This modification waits on the variable
to be set before stopping wazo-calld.

Having to wait in the test also indicates that restart calld results in
undefined behavior in this scenario. If wazo-calld gets stopped before writing
the channel state variable for a given call then the collectd event will not be
published when the call is destroyed if the variable is there, the event will be
published as expected.