dial_mobile: save the PICKUPMARK
When want to be able to identify that this dial mobile calls should be
intercepted or not based on the PICKUP MARK of the callee

The function PJSIP_ENDPOINT(<endpoint>,PICKUPMARK) is used because the mobile
being called does not have a channel to get the variable from at this point.

Saving the PICKUPMARK at this point avoids having to iterate through all PJSIP
endpoint to find a matching PICKUPMARK at the moment of the pickup. Iterating
through current dial mobile instances make for a way shorter list